+++ packages/insomnia-inso/package.json
@@ -64,7 +64,7 @@
     "enquirer": "^2.3.6",
     "insomnia-plugin-core-themes": "^3.19.0-beta.4",
     "insomnia-plugin-default-headers": "^3.19.0-beta.4",
-    "insomnia-plugin-request": "^3.19.0-beta.4",
+    "insomnia": "^3.19.0-beta.4",
     "insomnia-send-request": "^3.19.0-beta.4",
     "insomnia-testing": "^3.19.0-beta.4",
     "nedb": "^1.8.0",
