+++ packages/insomnia/config/config.json
@@ -17,7 +17,6 @@
   },
   "sentryDsn": "https://aaec2e800e644070a8daba5b7ad02c16@o1147619.ingest.sentry.io/6311804",
   "plugins": [
-    "insomnia-plugin-request",
     "insomnia-plugin-core-themes",
     "insomnia-plugin-default-headers"
   ]
