--- plugins/insomnia-plugin-request/README.md
@@ -1,3 +1,0 @@
-# Insomnia Request Template Tag
-
-This is a core Insomnia plugin.
