--- plugins/insomnia-plugin-request/package.json
@@ -1,27 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-plugin-request",
-  "version": "3.19.0-beta.4",
-  "author": "Kong <office@konghq.com>",
-  "description": "Insomnia request template tag",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "plugins/insomnia-plugin-request"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "index.js",
-  "insomnia": {
-    "name": "request",
-    "unlisted": true
-  },
-  "scripts": {
-    "test": "jest --silent"
-  },
-  "dependencies": {
-    "tough-cookie": "^4.1.3"
-  }
-}
